quic: implement frame creation functions
This patch implements the frame creation operations in struct
quic_frame_ops, including quic_frame_stream_append(), which attempts
to append a stream message to a stream frame.

These functions are used for creating frames in the transmit path. For
example, STREAM frames are created via .create() and enqueued on the
stream_list, then transmitted and moved to the transmitted_list.

If a packet is detected as lost, the frames are moved back to stream_list
for retransmission.

Data path: socket -> outqueue -> [frame] -> packet -> UDP tunnel

Frame encoding relies on the quic_put_xxx() APIs previously introduced in
the common utilities patch.

Signed-off-by: Pengtao He <hepengtao@xiaomi.com>
Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>
Signed-off-by: Xin Long <lucien.xin@gmail.com>