quic: support bind/listen/connect/accept/close()
This patch adds support for bind(), listen(), connect(), accept(), poll(),
close(), and shutdown() on QUIC sockets.

1. bind()

This patch implements quic_bind() to allow applications to bind a QUIC
endpoint to a local address and port and associate it with a lower-layer
UDP (tunnel) socket.

Applications use bind() to specify with which local address and port the
QUIC endpoint should associate itself.

The function prototype of bind() is

  int bind(int sd,
           struct sockaddr *addr,
           socklen_t addrlen);

and the arguments are

  sd: The socket descriptor returned by socket().
  addr: The address structure (struct sockaddr_in for an IPv4 address or
    struct sockaddr_in6 for an IPv6 address).
  addrlen: The size of the address structure.

bind() returns 0 on success and -1 in case of an error.

Applications cannot call bind() multiple times to associate multiple
addresses with an endpoint. After the first bind() call, all subsequent
calls will return an error. However, multiple applications may bind() to
the same address and port, sharing the same lower UDP socket in the kernel.

The IP address part of addr can be specified as a wildcard (e.g.,
INADDR_ANY for IPv4 or IN6ADDR_ANY_INIT or in6addr_any for IPv6).
If the IPv4 sin_port or IPv6 sin6_port is set to 0, the operating
system will choose an ephemeral port for the endpoint.

If bind() is not explicitly called before connect() on the client, the
system will automatically determine a valid source address based on the
routing table and assign an ephemeral port to bind the socket during
connect() (added later).

Completing the bind() process does not permit the QUIC endpoint to accept
inbound QUIC connection requests on the server. This capability is only
enabled after a listen() system call (added earlier), is performed on the
socket.

2. listen()

This patch implements quic_inet_listen() to auto-bind if necessary,
generate and install connection IDs, derive initial keys, and register
the socket in the listen hash table.

An application uses listen() to mark a socket as being able to accept new
connections.

The function prototype is

  int listen(int sd,
             int backlog);

and the arguments are

  sd: The socket descriptor of the endpoint.
  backlog: If backlog is non-zero, enable listening, else disable
    listening.

listen() returns 0 on success and -1 in case of an error.

Since ALPNs are considered when inserting a listen socket into the
hash table in quic_hash(), this allows the kernel to parse ALPN from
a Client Initial packet and direct the incoming request accordingly to
different listening sockets bound to the same address and port. These
sockets may belong to different user processes or kernel threads.

The ALPNs for sockets are configured via a socket option (added later)
prior to calling listen(). If no ALPNs are set before listen(), the
listening socket will only be able to accept client connections that
do not specify any ALPN.

3. connect()

This patch implements quic_connect() to perform destination routing,
select a source address and port (auto-binding if bind() was not
called), set socket addresses, generate connection IDs, and install
initial keys required for encrypting handshake packets.

Application use connect() to get the socket ready for handshake.

The function prototype is

  int connect(int sd,
              const struct sockaddr *addr,
              socklen_t addrlen);

and the arguments are

  sd: The socket descriptor of the endpoint.
  addr: The address structure (struct sockaddr_in for an IPv4 address
    or struct sockaddr_in6 for an IPv6 address).
  addrlen: The size of the address structure.

connect() returns 0 on success and -1 on error.

connect() must be called before sending any handshake message.

4. accept()

This patch implements quic_accept() to allow applications to accept
incoming QUIC connection requests. The function blocks until a connection
request is available in the accept queue. It then removes the request,
allocates a new socket, and transfers transport parameters and
configuration from the listening socket.

The new socket is initialized via sk_prot->init() and sk_prot->bind(),
followed by connection ID generation and installation of initial keys
needed to decrypt Client Initial packets.

Applications use the accept() call to remove a QUIC connection request
from the accept queue of the endpoint. A new socket descriptor will be
returned from accept() to represent the newly formed connection request.

The function prototype is

  int accept(int sd,
             struct sockaddr *addr,
             socklen_t *addrlen);

and the arguments are

  sd: The socket descriptor of the endpoint.
  addr: The address structure (struct sockaddr_in for an IPv4 address or
    struct sockaddr_in6 for an IPv6 address).
  addrlen: The size of the address structure.

The function returns the socket descriptor for the newly formed connection
request on success and -1 in case of an error.

Note that the incoming Client Initial packet triggers the accept() call,
and the TLS message carried by the Client Initial packet will be queued
in the receive queue of the socket returned by accept(). This TLS message
will then be received and processed by userspace through the returned
socket, ensuring that the TLS handshake is completed in userspace.

5. poll()

Applications use poll() (as well as epoll or select) to monitor socket
readiness for reading, writing, error conditions, or connection closure.

Polling behavior is as follows:

  EPOLLIN / EPOLLRDNORM is returned when there is incoming application
  data or, in the case of a listening socket, pending connection requests.

  EPOLLOUT / EPOLLWRNORM is returned when the socket has sufficient
  buffer space for writing.

  EPOLLERR or EPOLLPRI is returned if the socket has encountered an
  error.

  On connection closure, most poll flags are typically returned to
  notify the application.

6. close()

Applications use close() to gracefully close down a connection.

The function prototype is

  int close(int sd);

and the arguments are

  sd: The socket descriptor of the connection to be closed.
  close() returns 0 on success and -1 in case of an error.

After an application calls close() on a socket descriptor, no further
socket operations will succeed on that descriptor.

close() will send a CLOSE frame to the peer. The close information can
be set via the CONNECTION_CLOSE socket option (added later) before
calling close().

7. shutdown()

QUIC differs from TCP in that it does not have half close semantics.

The function prototype is

  int shutdown(int sd,
               int how);

and the arguments are

  sd: The socket descriptor of the connection to be closed.

  how: Specifies the type of shutdown

    SHUT_RD: Disables further receive operations. All incoming data and
      connection requests will be discard quietly.
    SHUT_WR: Disables further send operations. It will send a CLOSE frame.
    SHUT_RDWR: Similar to SHUT_WR.

shutdown() returns 0 on success and -1 in case of an error.

The implementation is capable of unblocking sendmsg(), recvmsg(), and
accept() operations with SHUT_RDWR.

Signed-off-by: Pengtao He <hepengtao@xiaomi.com>
Signed-off-by: Moritz Buhl <mbuhl@openbsd.org>
Signed-off-by: Stefan Metzmacher <metze@samba.org>
Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>
Signed-off-by: Xin Long <lucien.xin@gmail.com>